+++ packages/storm-stack/jest.config.ts
@@ -1,3 +1,3 @@
-import { getJestConfig } from '@storm-software/testing-tools';
+import { getJestConfig } from "@storm-software/testing-tools";
 
-export default getJestConfig('packages/storm-stack', true, 'storm-stack');
+export default getJestConfig("packages/storm-stack", true, "storm-stack");